requirements.txt
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
cx-Oracle==5.2.1
decorator==4.0.11
Django==1.10.5
django-filter==1.0.1
django-simple-history==1.8.2
djangorestframework==3.5.4
enum34==1.1.6
ipdb==0.10.2
ipython==5.2.1
ipython-genutils==0.1.0
MySQL-python==1.2.5
olefile==0.44
pathlib2==2.2.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.0.0
prompt-toolkit==1.0.10
ptyprocess==0.5.1
Pygments==2.2.0
scandir==1.4
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.1
wcwidth==0.1.7
